requirements.txt
djangorestframework
django-cors-headers
django-cleanup
django-ckeditor
django-filter
django-nested-inline
django-resized
pillow
pytils
drf-yasg



